+++ apps/docs/scripts/lint/rules/admonitions-no-stacking.ts
@@ -1,5 +1,4 @@
 import { MdxJsxFlowElement } from 'mdast-util-mdx'
-import { NODE_TYPE_FOR_RULE } from '../utils/symbols'
 
 function admonitionsNoStacking(content: MdxJsxFlowElement) {
   if (content.name !== 'Admonition') {
@@ -8,7 +7,5 @@ function admonitionsNoStacking(content: MdxJsxFlowElement) {
 
   console.log(JSON.stringify(content, null, 2))
 }
-
-admonitionsNoStacking[NODE_TYPE_FOR_RULE] = 'mdxJsxFlowElement'
 
 export { admonitionsNoStacking }
