--- apps/docs/scripts/lint/utils/symbols.ts
@@ -1,1 +1,0 @@
-export const NODE_TYPE_FOR_RULE = Symbol('Node types that this rule applies to')
